+++ components/landing/ProductItem.tsx
@@ -10,7 +10,6 @@ import {
 import { SingleProductProps } from "../../utils/types/landingpage";
 import {
   CartItemProps,
-  CompareProductProps,
   RootAppStateProps,
 } from "../../utils/types/reduxTypes";
 
@@ -59,29 +58,47 @@ const ProductItem = ({ productItem }: { productItem: SingleProductProps }) => {
   //   dispatch(addToCartAction(cartItem));
   // };
   const handleAddToComparelist = () => {
-    const compareItem: CompareProductProps = {
+    const compareItem: SingleProductProps = {
       id: productItem.id,
       thumbnail: productItem.thumbnail,
       title: productItem.title,
+      slug: productItem.slug,
+      status: productItem.status,
+      vendor: productItem.vendor,
+      category: productItem.category,
+      sub_category: productItem.sub_category,
+      sub_sub_category: productItem.sub_sub_category,
+      unit: productItem.unit,
       unit_price: productItem.unit_price,
+      purchase_price: productItem.purchase_price,
+      tax_in_percent: productItem.tax_in_percent,
+      discount_type: productItem.discount_type,
+      discount_amount: productItem.discount_amount,
+      shipping_time: productItem.shipping_time,
+      youtube_link: productItem.youtube_link,
+      product_tags: productItem.product_tags,
+      product_reviews: productItem.product_reviews,
       total_quantity: productItem.total_quantity,
-      // old_price: productItem.total_quantity,
+      // old_price: product.price,
       short_description: productItem.short_description,
-      // rating: "4",
-      // is_featured: false,
-      category_name: productItem.category_name ? productItem.category_name : "",
+      // rating: product.rating,
+      // is_featured: product.is_featured,
+      category_name: productItem.category_name,
       brand: productItem.brand,
       full_description: productItem.full_description,
       warranty: productItem.warranty,
-      // variation: "productItem.variation",
+      total_shipping_cost: productItem.total_shipping_cost,
     };
-    if (compareProducts.length < 4) {
-      dispatch(addToCompareAction(compareItem));
-    } else if (compareProducts.length === 4) {
-      dispatch(removeFromCompareAction(compareProducts[0].id));
-      dispatch(addToCompareAction(compareItem));
+    if(compareProducts){
+      if (compareProducts.length < 4) {
+        dispatch(addToCompareAction(compareItem));
+      } else if (compareProducts.length === 4) {
+        dispatch(removeFromCompareAction(compareProducts[0].id));
+        dispatch(addToCompareAction(compareItem));
+      }
     }
   };
+
   const handleWishlist = () => {
     if (wishlist?.length !== 0) {
       if (wishlist) {
